Corrected some compiler warning, worked around a bug preventing the unittest from running, and fixed a couple of bugs in the unittest (not OM itself) to get it to pass.
git-svn-id: https://openmalaria.googlecode.com/svn/trunk@1844 aa1ca794-fd11-11dd-ac65-f74f249cfaac

+++ include/WithinHost/WithinHostModel.h
@@ -118,7 +118,7 @@ public:
    * infections to baloon stupidly.
    * 
    * Exact constraint is: _MOI <= MAX_INFECTIONS. */
-  static const size_t MAX_INFECTIONS = 21;
+  static const int MAX_INFECTIONS = 21;
   
 protected:
   /** Updates for the immunity model − assumes _cumulativeh and _cumulativeY
